requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
pydantic==2.5.0
requests==2.31.0
google-genai==0.2.2
python-dotenv==1.0.0
gunicorn==21.2.0

